move the functions that actually call APIs and parse, to lib/helpers
