+++ lib/helpers/getLowestPricedOffers.js
@@ -1,6 +1,6 @@
-const { parseEndpoint } = require('./parseEndpoint');
+const { parseEndpoint } = require('../parseEndpoint');
 
-const parseLowestPricedOffers = require('./parsers/lowestPricedOffers');
+const parseLowestPricedOffers = require('../parsers/lowestPricedOffers');
 
 // TODO: perhaps we should merge ForASIN and ForSKU varities into a single call, with either
 // a parameter to say which to get, or intelligent determination based on the contents? ... probably
