+++ lib/helpers/getMarketplaces.js
@@ -1,5 +1,5 @@
-const { parseEndpoint } = require('./parseEndpoint');
-const parseMarketplaceData = require('./parsers/marketplaceData');
+const { parseEndpoint } = require('../parseEndpoint');
+const parseMarketplaceData = require('../parsers/marketplaceData');
 
 // TODO: upgrade to call ListMarketplaceParticipationsByNextToken when a NextToken
 // response is returned.  This seems unlikely that we'd ever be able to test this in reality, as
