+++ lib/helpers/getMatchingProduct.js
@@ -1,7 +1,7 @@
-const { parseEndpoint } = require('./parseEndpoint');
-const errors = require('./errors');
+const { parseEndpoint } = require('../parseEndpoint');
+const errors = require('../errors');
 
-const parseMatchingProduct = require('./parsers/matchingProduct');
+const parseMatchingProduct = require('../parsers/matchingProduct');
 
 const inputParser = opt => ({
     MarketplaceId: opt.marketplaceId || opt.MarketplaceId,
