+++ lib/helpers/getMyFeesEstimate.js
@@ -1,6 +1,6 @@
-const { parseEndpoint } = require('./parseEndpoint');
-const parseFeesEstimate = require('./parsers/feesEstimate');
-const { forceArray } = require('./util/transformers');
+const { parseEndpoint } = require('../parseEndpoint');
+const parseFeesEstimate = require('../parsers/feesEstimate');
+const { forceArray } = require('../util/transformers');
 
 /**
  * @typedef EstimateRequest - an Estimate Request object
