+++ lib/helpers/getOrder.js
@@ -1,4 +1,4 @@
-const { parseEndpoint } = require('./parseEndpoint');
+const { parseEndpoint } = require('../parseEndpoint');
 
 /**
  * Return orders by ID, i.e. Amazon Order ID.
