+++ lib/helpers/getProductCategories.js
@@ -1,4 +1,4 @@
-const { callEndpoint } = require('./callEndpoint');
+const { callEndpoint } = require('../callEndpoint');
 
 /**
  * call GetProductCategoriesForASIN for a single asin.
