+++ lib/helpers/listFinancialEvents.js
@@ -1,5 +1,5 @@
-const { parseEndpoint } = require('./parseEndpoint');
-const { forceArray } = require('./util/transformers');
+const { parseEndpoint } = require('../parseEndpoint');
+const { forceArray } = require('../util/transformers');
 
 // TODO: probably needs to handle nextToken
 // TODO: write some tests to more completely test this function's output
