+++ lib/helpers/listMatchingProducts.js
@@ -1,5 +1,5 @@
-const { parseEndpoint } = require('./parseEndpoint');
-const parseMatchingProduct = require('./parsers/matchingProduct');
+const { parseEndpoint } = require('../parseEndpoint');
+const parseMatchingProduct = require('../parsers/matchingProduct');
 
 const inputParser = opt => ({
     MarketplaceId: opt.marketplaceId || opt.MarketplaceId,
