+++ lib/helpers/listOrderItems.js
@@ -1,8 +1,8 @@
 /**
  * @module mws-advanced
  */
-const { parseEndpoint } = require('./parseEndpoint');
-const parseOrderItems = require('./parsers/orderItems');
+const { parseEndpoint } = require('../parseEndpoint');
+const parseOrderItems = require('../parsers/orderItems');
 
 // TODO: make a typedef for OrderItem and document it correctly
 /**
