+++ lib/helpers/listOrders.js
@@ -1,4 +1,4 @@
-const { parseEndpoint } = require('./parseEndpoint');
+const { parseEndpoint } = require('../parseEndpoint');
 
 // returns
 /*
