+++ lib/helpers/reports.js
@@ -5,7 +5,7 @@
 const { promisify } = require('util');
 const fs = require('fs');
 
-const errors = require('./errors');
+const errors = require('../errors');
 
 /**
  * Promisified version of fs.writeFile
@@ -17,8 +17,8 @@ const errors = require('./errors');
  */
 const writeFile = promisify(fs.writeFile);
 
-const { parseEndpoint } = require('./parseEndpoint');
-const sleep = require('./util/sleep');
+const { parseEndpoint } = require('../parseEndpoint');
+const sleep = require('../util/sleep');
 
 /**
  * @typedef ReportRequestInfo
