+++ lib/index.js
@@ -8,28 +8,28 @@
 // http://s3.amazonaws.com/devo.docs.developer.amazonservices.com/en_US/sellers/Sellers_ListMarketplaceParticipations.html
 
 const { parseEndpoint } = require('./parseEndpoint');
-const { getMarketplaces } = require('./get-marketplaces');
-const { listOrderItems } = require('./list-order-items');
-const { listOrders } = require('./list-orders');
-const { listFinancialEvents } = require('./list-financial-events');
-const { listInventorySupply } = require('./list-inventory-supply');
-const { getMatchingProductForId } = require('./get-matching-product');
+const { getMarketplaces } = require('./helpers/getMarketplaces');
+const { listOrderItems } = require('./helpers/listOrderItems');
+const { listOrders } = require('./helpers/listOrders');
+const { listFinancialEvents } = require('./helpers/listFinancialEvents');
+const { listInventorySupply } = require('./helpers/listInventorySupply');
+const { getMatchingProductForId } = require('./helpers/getMatchingProduct');
 
 const {
     getLowestPricedOffersForASIN,
     getLowestPricedOffersForSKU,
-} = require('./get-lowest-priced-offers');
+} = require('./helpers/getLowestPricedOffers');
 
 const {
     getProductCategoriesForAsins,
     getProductCategoriesForSkus,
-} = require('./get-product-categories');
+} = require('./helpers/getProductCategories');
 
-const { getMyFeesEstimate } = require('./get-my-fees-estimate');
-const { listMatchingProducts } = require('./list-matching-products');
+const { getMyFeesEstimate } = require('./helpers/getMyFeesEstimate');
+const { listMatchingProducts } = require('./helpers/listMatchingProducts');
 const constants = require('./constants');
 
-const { getOrder } = require('./get-order');
+const { getOrder } = require('./helpers/getOrder');
 
 const {
     getReport,
@@ -42,7 +42,7 @@ const {
     requestReport,
     manageReportSchedule,
     updateReportAcknowledgements,
-} = require('./reports');
+} = require('./helpers/reports');
 
 const MWSAdvanced = require('./MwsAdvanced');
